Refactored main activity to use MVP pattern. Closes #14
